requirements.txt
python-jobspy
pandas
google-genai
python-dotenv